requirements.txt
streamlit
pandas
python-dotenv
openai>=1.3.0
